extract interpolation functions
